- minor fix matrix type conversion (for linux)
git-svn-id: https://vmmlib.svn.sourceforge.net/svnroot/vmmlib/trunk@333 4d770e62-bb1e-0410-ad62-f2382d3e8790

+++ tests/matrix_test.cpp
@@ -891,8 +891,8 @@ matrix_test::run()
 	{
 		//matrix type conversion
 		matrix< 2, 3, float_t > matrix_type_a;
-		matrix< 2, 3, uint16_t > matrix_type_b;
-		matrix< 2, 3, uint16_t > matrix_type_b_check;
+		matrix< 2, 3, int > matrix_type_b;
+		matrix< 2, 3, int > matrix_type_b_check;
 		
 		matrix_type_a.fill(2.4); 
 		matrix_type_b_check.fill(2);
